Add a new data processing module
